requirements.txt
streamlit
pandas
plotly
requests
python-dotenv
streamlit-option-menu
